Move plugin-package-build behaviour into plugin-build-library
Rollup config is now handeld by plugin-build-library.

plugin-package-build has been emptied and deprecated in favour of using
plugin-build-library
