+++ packages/loom-plugin-build-library/package.json
@@ -32,25 +32,31 @@
   ],
   "dependencies": {
     "@jest/transform": "^27.0.2",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-commonjs": "^19.0.0",
     "@rollup/plugin-graphql": "^1.0.0",
     "@rollup/plugin-image": "^2.1.1",
     "@rollup/plugin-json": "^4.1.0",
+    "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/pluginutils": "^4.1.0",
     "@shopify/babel-preset": "^24.1.4",
     "@shopify/loom": "^0.7.0",
     "@shopify/loom-plugin-babel": "^0.3.0",
     "@shopify/loom-plugin-jest": "^0.6.0",
-    "@shopify/loom-plugin-package-build": "^0.7.0",
+    "@shopify/loom-plugin-rollup": "^0.4.0",
     "@shopify/loom-plugin-typescript": "^0.8.0",
     "@shopify/postcss-plugin": "^3.0.1",
     "@svgr/rollup": "^4.3.3",
+    "common-ancestor-path": "^1.0.1",
     "graphql-mini-transforms": "^3.2.0",
     "graphql-typescript-definitions": "^2.0.4",
     "node-sass": "^4.12.0",
     "postcss": "^7.0.36",
-    "postcss-modules": "^3.0.0"
+    "postcss-modules": "^3.0.0",
+    "rollup-plugin-node-externals": "^2.2.0"
   },
   "devDependencies": {
+    "@types/common-ancestor-path": "^1.0.0",
     "@types/node-sass": "^4.11.2",
     "@types/svgr__rollup": "^4.3.1"
   }
