+++ packages/loom-plugin-build-library/src/index.ts
@@ -1,3 +1,4 @@
 export {buildLibrary, buildLibraryWorkspace} from './plugin-build-library';
 
 export {babel} from '@shopify/loom-plugin-babel';
+export {rollupPlugins} from '@shopify/loom-plugin-rollup';
