+++ packages/loom-plugin-build-library/src/plugin-write-binaries.ts
