+++ packages/loom-plugin-build-library/src/plugin-write-entrypoints.ts
