+++ packages/loom-plugin-package-build/package.json
@@ -28,24 +28,5 @@
     "!build/ts/**/tests/",
     "index.js"
   ],
-  "dependencies": {
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^19.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@shopify/loom": "^0.7.0",
-    "@shopify/loom-plugin-rollup": "^0.4.0",
-    "common-ancestor-path": "^1.0.1",
-    "rollup-plugin-node-externals": "^2.2.0"
-  },
-  "peerDependencies": {
-    "@shopify/loom-plugin-babel": "^0.0.1"
-  },
-  "peerDependenciesMeta": {
-    "@shopify/loom-plugin-babel": {
-      "optional": true
-    }
-  },
-  "devDependencies": {
-    "@types/common-ancestor-path": "^1.0.0"
-  }
+  "dependencies": {}
 }
