--- packages/loom-plugin-package-build/src/index.ts
@@ -1,4 +1,0 @@
-export {packageBuild} from './plugin-package-build';
-export type {PackageBuildOptions} from './plugin-package-build';
-
-export {rollupPlugins} from '@shopify/loom-plugin-rollup';
