--- packages/loom-plugin-package-build/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../config/typescript/tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "build/ts",
-    "baseUrl": "src",
-    "rootDir": "src"
-  },
-  "include": ["src/**/*"],
-  "references": [
-    {"path": "../../tests"},
-    {"path": "../loom"},
-    {"path": "../loom-plugin-babel"},
-    {"path": "../loom-plugin-rollup"}
-  ]
-}
